ai: renamed a few files to make it more usable for ai tools
